+++ TabLayout/TabLayout.js
@@ -13,6 +13,7 @@ import {forward, forwardCustom, forwardWithPrevent, forProp, handle, not} from '
 import {is} from '@enact/core/keymap';
 import kind from '@enact/core/kind';
 import {cap, mapAndFilterChildren} from '@enact/core/util';
+import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
 import Spotlight, {getDirection} from '@enact/spotlight';
 import {getTargetByDirectionFromElement} from '@enact/spotlight/src/target';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
@@ -196,6 +197,14 @@ const TabLayoutBase = kind({
 		orientation: PropTypes.oneOf(['horizontal', 'vertical']),
 
 		/**
+		 * Indicates the content's text direction is right-to-left.
+		 *
+		 * @type {Boolean}
+		 * @private
+		 */
+		rtl: PropTypes.bool,
+
+		/**
 		 * Assign a custom size to horizontal tabs.
 		 *
 		 * Tabs in the horizontal orientation automatically stretch to fill the available width.
@@ -253,7 +262,6 @@ const TabLayoutBase = kind({
 			if (forwardWithPrevent('onKeyDown', ev, props) && direction && collapsed && orientation === 'vertical' && document.querySelector(`[data-spotlight-id='${spotlightId}']`).contains(target) && target.tagName !== 'INPUT') {
 				Spotlight.setPointerMode(false);
 				ev.preventDefault();
-
 				if (Spotlight.move(direction)) {
 					ev.stopPropagation();
 				} else if (document.querySelector(`[data-spotlight-id='${spotlightId}'] .${componentCss.content}`).contains(target)) {
@@ -265,11 +273,13 @@ const TabLayoutBase = kind({
 						forward('onExpand', ev, props);
 					}
 				}
+			} else if (is('enter')(keyCode) && !collapsed && document.querySelector(`[data-spotlight-id='${spotlightId}-tabs-expanded']`).contains(target) && target.tagName !== 'INPUT') {
+				ev.stopPropagation();
 			}
 		},
 		onKeyUp: (ev, props) => {
 			const {keyCode, target} = ev;
-			const {collapsed, 'data-spotlight-id': spotlightId, type} = props;
+			const {anchorTo, collapsed, orientation, 'data-spotlight-id': spotlightId, rtl, type} = props;
 			const popupPanelRef = document.querySelector(`[data-spotlight-id='${spotlightId}'] .${popupTabLayoutComponentCss.panel}`);
 
 			if (forwardWithPrevent('onKeyUp', ev, props) && type === 'popup' && is('cancel')(keyCode) && popupPanelRef?.contains(target) && popupPanelRef?.dataset.index === '0') {
@@ -278,6 +288,32 @@ const TabLayoutBase = kind({
 				}
 				Spotlight.move('left');
 				ev.stopPropagation();
+			} else if (is('enter')(keyCode) && !collapsed && document.querySelector(`[data-spotlight-id='${spotlightId}-tabs-expanded']`).contains(target) && target.tagName !== 'INPUT') {
+				Spotlight.setPointerMode(false);
+
+				let moveTo;
+				if (orientation === 'vertical') {
+					if (anchorTo === 'left') {
+						moveTo = 'right';
+					} else if (anchorTo === 'right') {
+						moveTo = 'left';
+					} else if (anchorTo === 'start') {
+						if (rtl) {
+							moveTo = 'left';
+						} else {
+							moveTo = 'right';
+						}
+					} else if (anchorTo === 'end') {
+						if (!rtl) {
+							moveTo = 'left';
+						} else {
+							moveTo = 'right';
+						}
+					}
+				} else {
+					moveTo = 'down';
+				}
+				Spotlight.move(moveTo);
 			}
 		},
 		onSelect: handle(
@@ -345,6 +381,7 @@ const TabLayoutBase = kind({
 		delete rest.anchorTo;
 		delete rest.onExpand;
 		delete rest.onTabAnimationEnd;
+		delete rest.rtl;
 
 		const contentSize = (collapsed ? dimensions.content.expanded : dimensions.content.normal);
 		const isVertical = orientation === 'vertical';
@@ -414,7 +451,8 @@ const TabLayoutDecorator = compose(
 		enterTo: 'last-focused',
 		// favor the content when collapsed and the tabs otherwise
 		defaultElement: [`.${componentCss.horizontal} .${componentCss.tabs} *`, `.${componentCss.collapsed} .${componentCss.content} *`, `.${componentCss.tabsExpanded} *`]
-	})
+	}),
+	I18nContextDecorator({rtlProp: 'rtl'})
 );
 
 // Currently not documenting the base output since it's not exported