+++ WizardPanels/WizardPanels.js
@@ -1,10 +1,11 @@
-import handle, {forProp, forwardCustomWithPrevent, not} from '@enact/core/handle';
+import handle, {forProp, forwardCustom, forwardCustomWithPrevent, not} from '@enact/core/handle';
 import kind from '@enact/core/kind';
 import EnactPropTypes from '@enact/core/internal/prop-types';
 import useChainRefs from '@enact/core/useChainRefs';
 import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
+import Spotlight from '@enact/spotlight';
 import SpotlightContainerDecorator, {spotlightDefaultClass} from '@enact/spotlight/SpotlightContainerDecorator';
-import {Column, Cell} from '@enact/ui/Layout';
+import {Cell, Column, Row} from '@enact/ui/Layout';
 import Changeable from '@enact/ui/Changeable';
 import ForwardRef from '@enact/ui/ForwardRef';
 import ViewManager from '@enact/ui/ViewManager';
@@ -14,6 +15,7 @@ import compose from 'ramda/src/compose';
 import {createContext, useRef, useState, useCallback, Children} from 'react';
 
 import $L from '../internal/$L';
+import Button from '../Button';
 import {Header} from '../Panels';
 import {PanelBase} from '../Panels/Panel';
 import {BasicArranger, CrossFadeArranger, CancelDecorator, FloatingLayerIdProvider, NavigationButton, useAutoFocus} from '../internal/Panels';
@@ -67,6 +69,15 @@ const WizardPanelsBase = kind({
 		'aria-label': PropTypes.string,
 
 		/**
+		 * Hint string read when focusing the close button.
+		 *
+		 * @type {String}
+		 * @default 'Exit quick guide'
+		 * @private
+		 */
+		closeButtonAriaLabel: PropTypes.string,
+
+		/**
 		 * Obtains a reference to the root node.
 		 *
 		 * @type {Function|Object}
@@ -94,6 +105,14 @@ const WizardPanelsBase = kind({
 		* @public
 		*/
 		footer: PropTypes.node,
+
+		/**
+		* When `true`, indicates full screen size content
+		*
+		* @type {Boolean}
+		* @private
+		*/
+		fullScreenContent: PropTypes.bool,
 
 		/**
 		* The currently selected panel.
@@ -172,6 +191,14 @@ const WizardPanelsBase = kind({
 		* @public
 		*/
 		onChange: PropTypes.func,
+
+		/**
+		 * Called when the close button is clicked.
+		 *
+		 * @type {Function}
+		 * @private
+		 */
+		onClose: PropTypes.func,
 
 		/**
 		 * Called when the next button is clicked in WizardPanel.
@@ -299,9 +326,13 @@ const WizardPanelsBase = kind({
 	},
 
 	handlers: {
+		onClose: forwardCustom('onClose'),
 		onNextClick: handle(
 			forwardCustomWithPrevent('onNextClick'),
-			(ev, {index, onChange, totalPanels}) => {
+			(ev, {'data-spotlight-id': spotlightId, fullScreenContent, index, onChange, totalPanels}) => {
+				if (fullScreenContent) {
+					Spotlight.set(spotlightId, {enterTo: 'last-focused'});
+				}
 				if (onChange && index !== totalPanels) {
 					const nextIndex = index < (totalPanels - 1) ? (index + 1) : index;
 
@@ -332,10 +363,11 @@ const WizardPanelsBase = kind({
 	},
 
 	computed: {
-		'aria-label': ({'aria-label': label, index, noSteps, subtitle, title}) => {
+		'aria-label': ({'aria-label': label, current, index, noSteps, subtitle, title}) => {
 			if (label) return label;
 
-			const step = noSteps ? '' : new IString($L('step {num}')).format({num: index + 1}) + ' ';
+			const stepNum = (typeof current === 'number' && current > 0) ? current : (index + 1);
+			const step = noSteps ? '' : new IString($L('step {num}')).format({num: stepNum}) + ' ';
 			return `${step}${title} ${subtitle}`;
 		},
 		className: ({noSteps, noSubtitle, styler}) => styler.append(
@@ -344,16 +376,76 @@ const WizardPanelsBase = kind({
 				noSubtitle
 			}
 		),
-		steps: ({current, index, noSteps, total, totalPanels}) => {
+		closeButton: ({closeButtonAriaLabel, onClose, totalPanels}) => {
+			return (
+				totalPanels ? <Button
+					aria-label={closeButtonAriaLabel == null ? $L('Exit quick guide') : closeButtonAriaLabel}
+					className={css.close}
+					icon="closex"
+					onClick={onClose}
+					size="small"
+				/> : null
+			);
+		},
+		steps: ({current, fullScreenContent, index, noSteps, total, totalPanels}) => {
 			const currentStep = (noSteps && 1) || ((typeof current === 'number' && current > 0) ? current : (index + 1));
 			const totalSteps = (noSteps && 1) || ((typeof total === 'number' && total > 0) ? total : totalPanels);
 
-			return (
+			return ( fullScreenContent ?
+				<Steps
+					className={css.steps}
+					pastIcon={'circle'}
+					currentIcon={'circle'}
+					futureIcon={'circle'}
+					current={currentStep}
+					layout="quickGuidePanels"
+					total={totalSteps}
+				/> :
 				<Steps
 					className={css.steps}
 					current={currentStep}
 					slot="slotAbove"
 					total={totalSteps}
+				/>
+			);
+		},
+		prevNavigationButton: ({fullScreenContent, index, onPrevClick, prevButton, prevButtonVisibility}) => {
+			const isPrevButtonVisible = prevButtonVisibility === 'always' || (prevButtonVisibility === 'auto' && index !== 0);
+
+			return (
+				<NavigationButton
+					aria-label={$L('Previous')}
+					backgroundOpacity="transparent"
+					component={prevButton}
+					focusEffectIconOnly
+					icon="arrowlargeleft"
+					className={css.navigationButton}
+					iconFlip="auto"
+					minWidth={false}
+					onClick={onPrevClick}
+					slot={fullScreenContent ? null : 'slotBefore'}
+					visible={isPrevButtonVisible}
+				/>
+			);
+		},
+		nextNavigationButton: ({fullScreenContent, index, nextButton, nextButtonVisibility, onNextClick, totalPanels}) => {
+			const isNextButtonVisible = nextButtonVisibility === 'always' || (nextButtonVisibility === 'auto' && index < totalPanels - 1);
+
+			return (
+				<NavigationButton
+					aria-label={$L('Next')}
+					backgroundOpacity="transparent"
+					component={nextButton}
+					focusEffectIconOnly
+					className={css.navigationButton}
+					icon="arrowlargeright"
+					iconFlip="auto"
+					iconPosition="after"
+					id={fullScreenContent ? 'fullScreenNextButton' : null}
+					minWidth={false}
+					onClick={onNextClick}
+					slot={fullScreenContent ? null : 'slotAfter'}
+					visible={isNextButtonVisible}
 				/>
 			);
 		}
@@ -362,33 +454,66 @@ const WizardPanelsBase = kind({
 	render: ({
 		'aria-label': ariaLabel,
 		children,
+		closeButton,
 		footer,
+		fullScreenContent,
 		index,
-		nextButton,
-		nextButtonVisibility,
+		prevNavigationButton,
+		nextNavigationButton,
 		noAnimation,
 		noSubtitle,
-		onNextClick,
-		onPrevClick,
 		onTransition,
 		onWillTransition,
-		prevButton,
-		prevButtonVisibility,
 		reverseTransition,
 		steps,
 		subtitle,
 		title,
-		totalPanels,
 		...rest
 	}) => {
-		delete rest.noSteps;
+		delete rest.closeButtonAriaLabel;
 		delete rest.current;
+		delete rest.nextButton;
+		delete rest.nextButtonVisibility;
+		delete rest.noSteps;
+		delete rest.onClose;
+		delete rest.onNextClick;
+		delete rest.onPrevClick;
+		delete rest.prevButton;
+		delete rest.prevButtonVisibility;
 		delete rest.total;
+		delete rest.totalPanels;
+		if (fullScreenContent) {
+			// eslint-disable-next-line enact/prop-types
+			delete rest.hideChildren;
+		}
 
-		const isPrevButtonVisible = prevButtonVisibility === 'always' || (prevButtonVisibility === 'auto' && index !== 0);
-		const isNextButtonVisible = nextButtonVisibility === 'always' || (nextButtonVisibility === 'auto' && index < totalPanels - 1);
-
-		return (
+		return ( fullScreenContent ?
+			<article role="region" aria-labelledby={`quickguidepanel_index_${index}`} ref={rest.componentRef}>
+				<Column aria-label={ariaLabel} id={`quickguidepanel_index_${index}`} {...rest}>
+					<Row className={css.fullScreenContentHeader}>
+						{steps}
+						{closeButton}
+					</Row>
+					<Row className={css.fullScreenNavigationButtonContainer}>
+						<Cell shrink>
+							{prevNavigationButton}
+						</Cell>
+						<Cell />
+						<Cell shrink>
+							{nextNavigationButton}
+						</Cell>
+					</Row>
+					<ViewManager
+						arranger={BasicArranger}
+						duration={400}
+						noAnimation
+						onTransition={onTransition}
+						onWillTransition={onWillTransition}
+					>
+						{children}
+					</ViewManager>
+				</Column>
+			</article> :
 			<DecoratedPanelBase
 				{...rest}
 				header={
@@ -404,31 +529,8 @@ const WizardPanelsBase = kind({
 						type="wizard"
 					>
 						{steps}
-						<NavigationButton
-							aria-label={$L('Previous')}
-							backgroundOpacity="transparent"
-							component={prevButton}
-							focusEffectIconOnly
-							icon="arrowlargeleft"
-							iconFlip="auto"
-							minWidth={false}
-							onClick={onPrevClick}
-							slot="slotBefore"
-							visible={isPrevButtonVisible}
-						/>
-						<NavigationButton
-							aria-label={$L('Next')}
-							backgroundOpacity="transparent"
-							component={nextButton}
-							focusEffectIconOnly
-							icon="arrowlargeright"
-							iconFlip="auto"
-							iconPosition="after"
-							minWidth={false}
-							onClick={onNextClick}
-							slot="slotAfter"
-							visible={isNextButtonVisible}
-						/>
+						{prevNavigationButton}
+						{nextNavigationButton}
 					</HeaderContainer>
 				}
 				panelType="wizard"
@@ -656,7 +758,7 @@ const WizardPanelsDecorator = compose(
 	SpotlightContainerDecorator({
 		continue5WayHold: true,
 		// prefer any spottable within the panel body (content or footer) followed by header
-		defaultElement: [`.${spotlightDefaultClass}`, `.${css.content} *, .${css.footer} *`, 'header > *'],
+		defaultElement: [`.${spotlightDefaultClass}`, `.${css.content} *, .${css.footer} *`, 'header > *', `#fullScreenNextButton`, `.${css.close} *`],
 		enterTo: 'default-element'
 	}),
 	I18nContextDecorator({rtlProp: 'rtl'}),